added abstract class for all components additional payment methods.
